[edit] 급여계산 소득세 수정, 사대보험 오타 수정

+++ orca/src/main/java/com/groupware/orca/salary/vo/ClientVo.java
@@ -7,7 +7,7 @@ public class ClientVo {
 
     private double holidayTime;
     private double overTime;
-    private double person;
+    private int person;
     private double position;
     private String empNo;
     private double bonus;
